pyproject.toml
[project]
name = "pneumoniaclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[gs]>=3.66.1",
    "django>=6.0.1",
    "evidently>=0.7.20",
    "fastapi>=0.116.1",
    "google-cloud-storage>=3.8.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pillow>=11.0.0",
    "prometheus-client>=0.24.1",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
    "uvicorn>=0.35.0",
    "wandb>=0.23.1",
    "bentoml>=1.4.33",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "coverage>=7.13.1",
    "httpx>=0.28.1",
    "invoke>=2.2.1",
    "locust>=2.38.0",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.setuptools]
package-dir = { "" = "src", "frontend_django" = "frontend_django" }
packages = ["pneumoniaclassifier", "frontend_django"]


[tool.ruff]
line-length = 120
select = [
    "ALL", # include all the rules, including new ones
]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos,
    "EM102"
]

[tool.ruff.per-file-ignores]
"tests/**" = ["S101"]  # assert is standard in pytest

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/config/*"]

[project.scripts]
train = "pneumoniaclassifier.train:app"
evaluate = "pneumoniaclassifier.evaluate:app"
data-check = "pneumoniaclassifier.data:app"
data-stats = "pneumoniaclassifier.data_statistics:app"
link-model = "pneumoniaclassifier.link_model:app"
serve-api = "pneumoniaclassifier.api:cli_app"
frontend = "frontend_django.cli_frontend:app"
